feat: Add Docker configuration files and scripts
This commit adds the necessary Docker configuration files and scripts to the repository. The `.dockerignore` file is created to specify which files and directories should be ignored by Docker. The `Dockerfile` is added to define the image configuration and dependencies. The `nginx-site.conf` file is included to configure the Nginx server. Finally, the `00-laravel-deploy.sh` script is added to handle the deployment of the Laravel application.