feat: Update database configuration and add UserSeeder
The database configuration file `database.php` is updated to use PostgreSQL as the default database connection. Additionally, the Redis prefix in the configuration is modified to include the application name.

The `DatabaseSeeder` class is modified to comment out the seeding of `User`, `Category`, and `Task` models.

The `00-laravel-deploy.sh` script is updated to include the creation of a `UserSeeder` using the `make:seeder` command.